Created an Interface for listening on UserStatus for login.
Created and added specifik icons and titles for each Startpage button.

Implemented two new classes, DataPair and DataNode for saving a map of DataNode-pairs for each button node, one for each UserStatus state. (Implemented in StartpageController.)

Re-visited Login/Register-forms, we need to make the forms Return-friendly and respond to these key-presses for posting the form to the database.